Use a ulong internally as popcnt works on uint64 as well
BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22631
AMD Ryzen 5 5600U with Radeon Graphics, 1 CPU, 12 logical and 6 physical cores
.NET SDK=8.0.300
  [Host]     : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT
  DefaultJob : .NET 8.0.5 (8.0.524.21615), X64 RyuJIT

Before:

|    Method |      Mean |     Error |    StdDev | Allocated |
|---------- |----------:|----------:|----------:|----------:|
| FirstTrue |  4.246 ns | 0.0769 ns | 0.0719 ns |         - |
| CountTrue | 11.362 ns | 0.2170 ns | 0.2412 ns |         - |
|      NAnd |  9.507 ns | 0.1762 ns | 0.1562 ns |         - |
|    From_S |  5.194 ns | 0.0821 ns | 0.0728 ns |         - |
|    From_M |  4.427 ns | 0.0493 ns | 0.0927 ns |         - |
|    From_L |  5.258 ns | 0.0472 ns | 0.0419 ns |         - |
|   From_XL | 11.686 ns | 0.0355 ns | 0.0315 ns |         - |
|  From_XXL | 57.214 ns | 0.2519 ns | 0.2233 ns |         - |
|     Xor_L | 17.547 ns | 0.5763 ns | 1.5481 ns |         - |

After:

|    Method |      Mean |     Error |    StdDev |    Median | Allocated |
|---------- |----------:|----------:|----------:|----------:|----------:|
| FirstTrue |  5.236 ns | 0.0350 ns | 0.0328 ns |  5.214 ns |         - |
| CountTrue |  8.840 ns | 0.4875 ns | 1.4375 ns |  8.251 ns |         - |
|      NAnd |  8.851 ns | 0.1573 ns | 0.2836 ns |  8.766 ns |         - |
|    From_S |  4.517 ns | 0.0674 ns | 0.0598 ns |  4.511 ns |         - |
|    From_M |  4.481 ns | 0.0427 ns | 0.0400 ns |  4.473 ns |         - |
|    From_L |  4.778 ns | 0.0849 ns | 0.0752 ns |  4.781 ns |         - |
|   From_XL | 11.928 ns | 0.1112 ns | 0.0868 ns | 11.897 ns |         - |
|  From_XXL | 57.220 ns | 0.4555 ns | 0.3803 ns | 57.131 ns |         - |
|     Xor_L |  8.549 ns | 0.1631 ns | 0.1602 ns |  8.491 ns |         - |

+++ src/MonoTorrent/MonoTorrent/BitField.cs
@@ -47,7 +47,7 @@ namespace MonoTorrent
 
         public double PercentComplete => Data.PercentComplete;
 
-        internal Span<uint> Span => Data.Data;
+        internal Span<ulong> Span => Data.Data;
 
         public int TrueCount => Data.TrueCount;
